+++ expensetrackerapi/Models/BucketTransaction.cs
@@ -1,4 +1,4 @@
 ﻿namespace expensetrackerapi.Models;
 
 
-public record BucketTransaction(int BucketId, string BucketName, BucketTypes BucketType, int BucketExpenseTotal, Transaction[] Transactions);
+public record BucketTransaction(int BucketId, string BucketName, BucketTypes BucketType, decimal BucketExpenseTotal, Transaction[] Transactions);
